requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
PyMySQL==1.1.0
Werkzeug==3.0.1
openai==1.54.0
